refactor: 임시 비밀번호 발급 페이지 도메인 변경(auth -> user)
