+++ src/features/user/ui/index.ts
@@ -1,3 +1,4 @@
 import My from './My';
+import PasswordRequestLayout from './PasswordRequestLayout';
 
-export { My };
+export { My, PasswordRequestLayout };
